split-utils * Move useCounter to hooks/useCounter.ts
