+++ src/containers/DashboardWrapper/DashboardWrapper.tsx
@@ -1,10 +1,11 @@
 import React from 'react'
 import { Loader } from '@entur/loader'
 import { Station, Vehicle } from '@entur/sdk/lib/mobility/types'
-import { useCounter, isDarkOrDefaultTheme } from '../../utils'
+import { isDarkOrDefaultTheme } from '../../utils'
 import { useSettings } from '../../settings/SettingsProvider'
 import { EnturLogo } from '../../assets/icons/EnturLogo'
 import { StopPlaceWithDepartures } from '../../types'
+import { useCounter } from '../../hooks/useCounter'
 import { ThemeContrastWrapper } from '../ThemeWrapper/ThemeContrastWrapper'
 import { NoStopsOnTavle } from '../Error/ErrorPages'
 import { BottomMenu } from './BottomMenu/BottomMenu'
