+++ src/dashboards/Timeline/TimelineDashboard.tsx
@@ -10,7 +10,6 @@ import {
     getIcon,
     getIconColor,
     timeUntil,
-    useCounter,
     getIconColorType,
     usePrevious,
     isEqualUnsorted,
@@ -30,6 +29,7 @@ import {
     saveToLocalStorage,
 } from '../../settings/LocalStorage'
 import { LongPressProvider } from '../../logic/longPressContext'
+import { useCounter } from '../../hooks/useCounter'
 import './TimelineDashboard.scss'
 
 const TICKS = [-1, 0, 1, 2, 3, 4, 5, 10, 15, 20, 30, 60]
